requirements.txt
httpx
pyyaml
mkdocs-material
shot-scraper
jinja2